Make library either static or shared.
Add some platform-specific compiler settings and claenup warnings. Add
option to control the type of library being built: STATIC or
SHARED. To handle this, add platform.h header to setup generic EXPORT
macros.

+++ include/json_spirit/value.h
@@ -331,7 +331,7 @@ namespace json_spirit
         {
             std::ostringstream os;
 
-            os << "value type is " << type() << " not " << vtype;
+            os << "value type is " << (int)type() << " not " << (int)vtype;
 
             throw std::runtime_error( os.str() );
         }
