+++ test/reader_test.cpp
@@ -434,7 +434,7 @@ namespace
 
             const Array_type arr = value.get_array();
 
-            assert_eq( arr.size(), 6 );
+            assert_eq( arr.size(), (unsigned int)6 );
             assert_eq( arr[0].get_real(), 1.200000000000000, 1e-16 );
             assert_eq( arr[1].get_real(), 1.234567890123456e+125, 1e+110 );
             assert_eq( arr[2].get_real(), -1.234000000000000e-123, 1e+108 );
